+++ src/components/PayloadSamples/styled.elements.ts
@@ -1,5 +1,6 @@
 // @ts-ignore
-import Dropdown from 'react-dropdown';
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import ReactDropdown from 'react-dropdown';
 
 import { transparentize } from 'polished';
 import styled from '../../styled-components';