+++ src/components/SearchBox/styled.elements.tsx
@@ -1,5 +1,6 @@
-import { darken } from 'polished';
 import * as React from 'react';
+
+import { darken, getLuminance, lighten } from 'polished';
 import styled from '../../styled-components';
 import { MenuItemLabel } from '../SideMenu/styled.elements';
 
@@ -16,9 +17,15 @@ export const SearchInput = styled.input.attrs(() => ({
   padding: 5px ${props => props.theme.spacing.unit * 2}px 5px
     ${props => props.theme.spacing.unit * 4}px;
   border: 0;
-  border-bottom: 1px solid ${({theme}) => darken(0.1, theme.menu.backgroundColor)};
-  font-family: ${({theme}) => theme.typography.fontFamily};
-  font-size: 1em;
+  border-bottom: 1px solid
+    ${({ theme }) =>
+      (getLuminance(theme.menu.backgroundColor) > 0.5 ? darken : lighten)(
+        0.1,
+        theme.menu.backgroundColor,
+      )};
+  font-family: ${({ theme }) => theme.typography.fontFamily};
+  font-weight: bold;
+  font-size: 13px;
   color: ${props => props.theme.menu.textColor};
   background-color: transparent;
   outline: none;
@@ -26,23 +33,14 @@ export const SearchInput = styled.input.attrs(() => ({
 
 export const SearchIcon = styled((props: { className?: string }) => (
   <svg
-    version="1.1"
-    id="Layer_1"
-    xmlns="http://www.w3.org/2000/svg"
-    x="0px"
-    y="0px"
     className={props.className}
-    viewBox="0 0 24 24"
-    xmlSpace="preserve"
+    version="1.1"
+    viewBox="0 0 1000 1000"
+    x="0px"
+    xmlns="http://www.w3.org/2000/svg"
+    y="0px"
   >
-    <g>
-      <path
-        className="st0"
-        d="M22.7,21.5l-5.1-5c1.5-1.7,2.4-4,2.4-6.5c0-5.5-4.5-10-10-10S0,4.5,0,10s4.5,10,10,10c2.3,0,4.4-0.8,6.1-2.1
-		l5.2,5.1c0.2,0.2,0.4,0.3,0.7,0.3c0.3,0,0.5-0.1,0.7-0.3C23.1,22.5,23.1,21.9,22.7,21.5z M10,18c-4.4,0-8-3.6-8-8s3.6-8,8-8
-		s8,3.6,8,8S14.4,18,10,18z"
-      />
-    </g>
+    <path d="M968.2,849.4L667.3,549c83.9-136.5,66.7-317.4-51.7-435.6C477.1-25,252.5-25,113.9,113.4c-138.5,138.3-138.5,362.6,0,501C219.2,730.1,413.2,743,547.6,666.5l301.9,301.4c43.6,43.6,76.9,14.9,104.2-12.4C981,928.3,1011.8,893,968.2,849.4z M524.5,522c-88.9,88.7-233,88.7-321.8,0c-88.9-88.7-88.9-232.6,0-321.3c88.9-88.7,233-88.7,321.8,0C613.4,289.4,613.4,433.3,524.5,522z" />
   </svg>
 )).attrs({
   className: 'search-icon',
@@ -59,11 +57,12 @@ export const SearchIcon = styled((props: { className?: string }) => (
 
 export const SearchResultsBox = styled.div`
   padding: ${props => props.theme.spacing.unit}px 0;
-  background-color: #ededed;
+  background-color: ${({ theme }) => darken(0.05, theme.menu.backgroundColor)}};
+  color: ${props => props.theme.menu.textColor};
   min-height: 150px;
   max-height: 250px;
-  border-top: 1px solid #e1e1e1;
-  border-bottom: 1px solid #e1e1e1;
+  border-top: ${({ theme }) => darken(0.1, theme.menu.backgroundColor)}};
+  border-bottom: ${({ theme }) => darken(0.1, theme.menu.backgroundColor)}};
   margin-top: 10px;
   line-height: 1.4;
   font-size: 0.9em;
@@ -72,16 +71,13 @@ export const SearchResultsBox = styled.div`
     padding-top: 6px;
     padding-bottom: 6px;
 
-    &:hover {
-      background-color: #e1e1e1;
+    &:hover,
+    &.active {
+      background-color: ${({ theme }) => darken(0.1, theme.menu.backgroundColor)};
     }
 
     > svg {
       display: none;
-    }
-
-    &.active {
-      background-color: #e1e1e1;
     }
   }
 `;