+++ src/components/SideMenu/SideMenu.tsx
@@ -1,12 +1,11 @@
 import { observer } from 'mobx-react';
 import * as React from 'react';
-import { StyledLink } from '../../../src/common-elements';
-
-import { PerfectScrollbarWrap } from '../../common-elements/perfect-scrollbar';
 
 import { IMenuItem, MenuStore } from '../../services/MenuStore';
 import { OptionsContext } from '../OptionsProvider';
 import { MenuItems } from './MenuItems';
+
+import { PerfectScrollbarWrap } from '../../common-elements/perfect-scrollbar';
 import { RedocAttribution } from './styled.elements';
 
 @observer
@@ -24,11 +23,11 @@ export class SideMenu extends React.Component<{ menu: MenuStore; className?: str
           wheelPropagation: false,
         }}
       >
-        <MenuItems items={store.items} onActivate={this.activate} root={true}/>
+        <MenuItems items={store.items} onActivate={this.activate} root={true} />
         <RedocAttribution>
-          <StyledLink href="https://www.opentext.com/" target={'_blank'}>
+          <a target="_blank" rel="noopener noreferrer" href="https://www.opentext.com/">
             © Copyright 2019 OpenText Corp
-          </StyledLink>
+          </a>
         </RedocAttribution>
       </PerfectScrollbarWrap>
     );