+++ src/utils/helpers.ts
@@ -189,7 +189,7 @@ export function removeQueryString(serverUrl: string): string {
 function parseURL(url: string) {
   if (typeof URL === 'undefined') {
     // node
-    return new (require('url')).URL(url);
+    return new (require('url').URL)(url);
   } else {
     return new URL(url);
   }